🛠️ Refactor: User, jwt Service Layer Pattern 적용
Related to: #29
